cleaning up comments and bad SVN tags;
git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project@952 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e

+++ trunk/modules/tasks/listtasks.php
@@ -7,7 +7,7 @@ $perms = &$AppUI->acl();
 if (!$perms->checkModule('tasks', 'view')) {
 	$AppUI->redirect('m=public&a=access_denied');
 }
-$proj = w2PgetParam($_GET, 'project', 0);
+$proj = (int) w2PgetParam($_GET, 'project', 0);
 $userFilter = w2PgetParam($_GET, 'userFilter', false);
 
 $q = new DBQuery();
